Move file commands into file/ subdirectory
Rename filecontent/filemanagement/fileoperation to
commands/file/content, commands/file/management, commands/file/operation.
