+++ include/core/commands/file/operation.hpp
@@ -1,5 +1,5 @@
 /**
- * filecontent.hpp
+ * operation.hpp
  *
  * Copyright (c) 2019-2026 Giovanni Giacomo. All Rights Reserved.
  * Use of this source code is governed by a MIT-style
@@ -7,31 +7,24 @@
  *
  */
 
-#ifndef CORE_COMMANDS_FILECONTENT_HPP_
-#define CORE_COMMANDS_FILECONTENT_HPP_
+#ifndef CORE_COMMANDS_FILE_OPERATION_HPP_
+#define CORE_COMMANDS_FILE_OPERATION_HPP_
 
 #include <core/commands/command.hpp>
 
 namespace cassio {
 namespace kernel {
 
-class CatCommand : public Command {
+class MvCommand : public Command {
 public:
-    CatCommand();
+    MvCommand();
     bool execute(const char** args, usize argc,
                  filesystem::FileNode*& cwd) override;
 };
 
-class WriteCommand : public Command {
+class CpCommand : public Command {
 public:
-    WriteCommand();
-    bool execute(const char** args, usize argc,
-                 filesystem::FileNode*& cwd) override;
-};
-
-class EchoCommand : public Command {
-public:
-    EchoCommand();
+    CpCommand();
     bool execute(const char** args, usize argc,
                  filesystem::FileNode*& cwd) override;
 };
@@ -39,4 +32,4 @@ public:
 } // kernel
 } // cassio
 
-#endif // CORE_COMMANDS_FILECONTENT_HPP_
+#endif // CORE_COMMANDS_FILE_OPERATION_HPP_
