+++ src/core/commands/file/content.cpp
@@ -1,5 +1,5 @@
 /**
- * filecontent.cpp
+ * content.cpp
  *
  * Copyright (c) 2019-2026 Giovanni Giacomo. All Rights Reserved.
  * Use of this source code is governed by a MIT-style
@@ -7,7 +7,7 @@
  *
  */
 
-#include "core/commands/filecontent.hpp"
+#include "core/commands/file/content.hpp"
 #include "common/string.hpp"
 #include "filesystem/filesystem.hpp"
 #include "hardware/terminal.hpp"
