+++ src/core/commands/file/operation.cpp
@@ -1,5 +1,5 @@
 /**
- * fileoperation.cpp
+ * operation.cpp
  *
  * Copyright (c) 2019-2026 Giovanni Giacomo. All Rights Reserved.
  * Use of this source code is governed by a MIT-style
@@ -7,7 +7,7 @@
  *
  */
 
-#include "core/commands/fileoperation.hpp"
+#include "core/commands/file/operation.hpp"
 #include "filesystem/filesystem.hpp"
 #include "hardware/terminal.hpp"
